+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Use deterministic deployments in migrations.
 - Update dependency version `@animoca/ethereum-contracts@0.0.5`.
+- Fix bug in `skipIfFileExists`.
 
 ## 0.0.1